Create and run configured test-suite standalone app (for simulator)
We want to test test-suite in a standalone app in CI. The first step of that is creating the standalone app. This adds a new job called `shell_app_ios_tests` that uses the published test-suite URL and configures an .app directory and runs it.

To make this a bit easier I use a new option called `--writeUrl` to `powertools publish-test-suite` to save the published project URL to a file that can be shared across jobs.

The second step is to run the configured standalonoe app using the new powertools command.

Test Plan: let CI run and make sure we can download the test-suite .app as an artifact and install and run it on a simulator locally.

fbshipit-source-id: eabfd65